Yank the Settler AI code out into its own class.  Gonna refactor it from here, but now it's in its own component.
Also add a fix in PlayerAI so it doesn't crash if an explorer AI has nowhere to move.  Not sure why this happens yet, and it's rare, but the new print statement (and ability to set a breakpoint) ought to help figure it out.
